Make SelfManagedApicast more flexible
Primary idea behind this change is:
 - get ability to set instance of SelfManagedApicast explicitly
   (force option)

Secondary benefit is:
 - (CHANGE IN BEHAVIOR!) more dynamic selection
   (fallback to TemplateApicast actually as this is "always" available)
