+++ testsuite/gateways/apicast/template.py
@@ -6,13 +6,12 @@ import importlib_resources as resources
 
 from testsuite.openshift.objects import SecretTypes
 from testsuite.openshift.client import OpenShiftClient
-from .selfmanaged import SelfManagedApicast
-from ...capabilities import CapabilityRegistry, Capability
+from . import OpenshiftApicast
 
 LOGGER = logging.getLogger(__name__)
 
 
-class TemplateApicast(SelfManagedApicast):
+class TemplateApicast(OpenshiftApicast):
     """Template-based APIcast Gateway."""
 
     # pylint: disable=too-many-arguments
@@ -37,8 +36,8 @@ class TemplateApicast(SelfManagedApicast):
                 "CONFIGURATION_CACHE": 0})
 
     @staticmethod
-    def fits():
-        return Capability.OCP3 in CapabilityRegistry()
+    def fits(openshift: OpenShiftClient):  # pylint: disable=unused-argument
+        return True
 
     def _create_configuration_url_secret(self):
         self.openshift.secrets.create(
